feat: supplement cf facade
