+++ packages/sheets-conditional-formatting-ui/src/controllers/cf.permission.controller.ts
@@ -15,10 +15,10 @@
  */
 
 import type { ICommandInfo } from '@univerjs/core';
-import type { IAddCfCommandParams } from '../commands/commands/add-cf.command';
+import type { IAddCfCommandParams } from '@univerjs/sheets-conditional-formatting';
 import { Disposable, ICommandService, Inject, LocaleService } from '@univerjs/core';
 import { RangeProtectionPermissionEditPoint, SheetPermissionCheckController, WorkbookEditablePermission, WorksheetEditPermission, WorksheetSetCellStylePermission } from '@univerjs/sheets';
-import { AddCfCommand } from '../commands/commands/add-cf.command';
+import { AddCfCommand } from '@univerjs/sheets-conditional-formatting';
 
 export class ConditionalFormattingPermissionController extends Disposable {
     constructor(
