+++ packages/sheets-conditional-formatting-ui/src/index.ts
@@ -23,7 +23,6 @@ export { ConditionalFormattingClearController } from './controllers/cf.clear.con
 // #region - all commands
 
 export { AddAverageCfCommand } from './commands/commands/add-average-cf.command';
-export { AddCfCommand, type IAddCfCommandParams } from './commands/commands/add-cf.command';
 export { AddColorScaleConditionalRuleCommand } from './commands/commands/add-color-scale-cf.command';
 export { AddDataBarConditionalRuleCommand } from './commands/commands/add-data-bar-cf.command';
 export { AddDuplicateValuesCfCommand } from './commands/commands/add-duplicate-values-cf.command';
@@ -32,11 +31,6 @@ export { AddRankCfCommand } from './commands/commands/add-rank-cf.command';
 export { AddTextCfCommand } from './commands/commands/add-text-cf.command';
 export { AddTimePeriodCfCommand } from './commands/commands/add-time-period-cf.command';
 export { AddUniqueValuesCfCommand } from './commands/commands/add-unique-values-cf.command';
-export { ClearRangeCfCommand } from './commands/commands/clear-range-cf.command';
-export { ClearWorksheetCfCommand } from './commands/commands/clear-worksheet-cf.command';
-export { DeleteCfCommand } from './commands/commands/delete-cf.command';
-export { MoveCfCommand } from './commands/commands/move-cf.command';
-export { SetCfCommand } from './commands/commands/set-cf.command';
 
 export { OpenConditionalFormattingOperator } from './commands/operations/open-conditional-formatting-panel';
