+++ packages/sheets-conditional-formatting/src/commands/commands/add-cf.command.ts
@@ -15,6 +15,9 @@
  */
 
 import type { ICommand } from '@univerjs/core';
+import type { IConditionFormattingRule } from '../../models/type';
+import type { MakePropertyOptional } from '../../utils/type';
+import type { IAddConditionalRuleMutationParams } from '../mutations/add-conditional-rule.mutation';
 import {
     CommandType,
     ICommandService,
@@ -22,8 +25,8 @@ import {
     IUniverInstanceService,
 } from '@univerjs/core';
 import { getSheetCommandTarget } from '@univerjs/sheets';
-import type { IAddConditionalRuleMutationParams, IConditionFormattingRule, MakePropertyOptional } from '@univerjs/sheets-conditional-formatting';
-import { AddConditionalRuleMutation, AddConditionalRuleMutationUndoFactory, ConditionalFormattingRuleModel } from '@univerjs/sheets-conditional-formatting';
+import { ConditionalFormattingRuleModel } from '../../models/conditional-formatting-rule-model';
+import { AddConditionalRuleMutation, AddConditionalRuleMutationUndoFactory } from '../mutations/add-conditional-rule.mutation';
 
 export interface IAddCfCommandParams {
     unitId?: string;
